Implement Worker.prototype.terminate()
Adds support for terminating DOM workers. A new ScriptMsg::Terminate was
added to notify the WorkerGlobalScope/DedicatedWorkerGlobalScope to stop
processing events and shut down. A closing flag was added to
WorkerGlobalScope per the spec.

Additionally, a closing flag was added to dom::worker::Worker to ensure
no further events are dispatched to the onmessage handler.

The change to the wpt WorkerTerminate.js is a patch for the broken
Worker_terminate_event_queue test (asserts 10001 == 10000).

Resolves servo/servo#4427